# artifacts/openapi.yaml
openapi: 3.0.3
info:
  title: Household Resource Planner API
  description: |
    API for tracking household consumables, forecasting depletion, and generating smart shopping lists.

    ## Authentication
    All endpoints require a valid JWT bearer token issued by Keycloak.

    ## Calculation Date
    All forecast, shopping-list, and dashboard operations are computed relative to a **calculationDate**
    variable stored server-side per user. By default it equals today's date.
    Use `GET /settings/calculation-date` to read it and `PUT /settings/calculation-date` to change it —
    for example to simulate a future date during a demo.
  version: 1.0.0
servers:
  - url: http://localhost:8081/api/v1
    description: Local dev (Docker backend on host port 8081)
  - url: http://localhost:8080/api/v1
    description: Local dev (backend run directly)
  - url: https://your-host/api/v1
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ConsumerCategory:
      type: string
      enum: [ADULT, CHILD, CAT, DOG, PARROT, SMALL_ANIMAL]
      description: Who this product is primarily consumed by
    ItemCategory:
      type: string
      enum:
        - FOOD
        - BEVERAGES
        - CLEANING
        - PERSONAL_CARE
        - PET_FOOD
        - MEDICINE
        - OTHER
...
    Item:
      type: object
      required: [name, currentQuantity, unit]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          example: Oat milk
        category:
          $ref: '#/components/schemas/ItemCategory'
        currentQuantity:
          type: number
          format: double
          description: Current amount in stock
          example: 2.5
        unit:
          type: string
          description: Unit of measurement (liters, kg, pcs, etc.)
          example: L
        storeId:
          type: string
          format: uuid
          description: Preferred store to buy this item
        price:
          type: number
          format: double
          example: 89.99
        consumerCategory:
          $ref: '#/components/schemas/ConsumerCategory'
        monthlyConsumptionRate:
          type: number
          format: double
          description: |
            Manual override for consumption rate (units per month).
            If null and autoCalc is true, the app infers it from purchase history.
          example: 8.0
        autoCalc:
          type: boolean
          default: false
          description: |
            Controls how the consumption rate is determined over time.
            - true + monthlyConsumptionRate set: use rate as initial value, then adjust based on real purchases
            - false + monthlyConsumptionRate set: always use the manually entered rate, never adjust
            - true + monthlyConsumptionRate null: infer rate from purchase history (no initial estimate)
        lastBoughtDate:
          type: string
          format: date
          description: Date when this item was last purchased. Used as the anchor for depletion forecasting.
          example: "2026-03-04"
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    CreateItemRequest:
      type: object
      required: [name, currentQuantity, unit]
      properties:
        name:
          type: string
        category:
          $ref: '#/components/schemas/ItemCategory'
        currentQuantity:
          type: number
          format: double
        unit:
          type: string
        storeId:
          type: string
          format: uuid
        price:
          type: number
          format: double
        consumerCategory:
          $ref: '#/components/schemas/ConsumerCategory'
        monthlyConsumptionRate:
          type: number
          format: double
        autoCalc:
          type: boolean
          default: false
        lastBoughtDate:
          type: string
          format: date
          description: Date when this item was last purchased. Defaults to today if omitted.
          example: "2026-03-04"
    UpdateItemRequest:
      allOf:
        - $ref: '#/components/schemas/CreateItemRequest'
    AddQuantityRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: number
          format: double
          description: Quantity to add to existing item
    SuggestMatchesRequest:
      type: object
      required: [productName]
      properties:
        productName:
          type: string
          description: Parsed product name from receipt/email
    MatchSuggestion:
      type: object
      properties:
        itemId:
          type: string
          format: uuid
        name:
          type: string
        score:
          type: number
          format: double
          description: Confidence 0-1
    MarkConsumedRequest:
      type: object
      description: Report that an item has been partially or fully used
      properties:
        quantityConsumed:
          type: number
          format: double
          description: How much was consumed. If omitted, item is marked as fully depleted.
          example: 1.0
        depletedAt:
          type: string
          format: date
          description: Date of consumption (defaults to today)
    SetConsumptionRateRequest:
      type: object
      required: [monthlyRate]
      properties:
        monthlyRate:
          type: number
          format: double
          description: Units of this item consumed per month
          example: 6.0

    # ─── Import ──────────────────────────────────────────────────────────────
...
paths:
...
  /items:
    get:
      tags: [Items]
      summary: List all tracked items
      operationId: listItems
      parameters:
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/ItemCategory'
        - name: storeId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
    post:
      tags: [Items]
      summary: Add a new item to track
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
  /items/suggest-matches:
    post:
      tags: [Items]
      summary: Suggest existing items that may match a parsed product name
      description: Uses LLM to match a product name from receipt/email to user's inventory.
      operationId: suggestMatches
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuggestMatchesRequest'
      responses:
        '200':
          description: Ranked list of possible matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchSuggestion'
  /items/{itemId}:
    get:
      tags: [Items]
      summary: Get an item
      operationId: getItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found
    put:
      tags: [Items]
      summary: Update an item
      operationId: updateItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemRequest'
      responses:
        '200':
          description: Updated item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found
    delete:
      tags: [Items]
      summary: Stop tracking an item
      operationId: deleteItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Item deleted
        '404':
          description: Item not found
  /items/{itemId}/add-quantity:
    post:
      tags: [Items]
      summary: Add quantity to existing item (restock)
      operationId: addQuantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddQuantityRequest'
      responses:
        '200':
          description: Updated item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found
  /items/{itemId}/consumed:
    post:
      tags: [Items]
      summary: Report consumption (or mark item as fully depleted)
      operationId: markConsumed
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkConsumedRequest'
      responses:
        '200':
          description: Updated item after consumption recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found
  /items/{itemId}/consumption-rate:
    put:
      tags: [Items]
      summary: Manually set monthly consumption rate for an item
      description: |
        Corresponds to user saying "I need X units of product Y per month".
        Overrides the app's automatic inference.
      operationId: setConsumptionRate
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetConsumptionRateRequest'
      responses:
        '200':
          description: Updated item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found

  # ── Import ─────────────────────────────────────────────────────────────────
...